pyproject.toml
[tool.poetry]
name = "script-sheets-python-docs"
version = "0.1.0"
description = ""
authors = ["fecaridade <caridade500@gmail.com>"]
license = "GPLV3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
